refactor: structure folder of my chat section
