+++ src/screens/chat/sections/my-chat/index.ts
@@ -1,0 +1,3 @@
+import MyChatSection from "./MyChatSection";
+
+export default MyChatSection;
